Refactor har-parser module into smaller, focused files
